remove unnecessary casts

+++ cockatrice/src/phasestoolbar.cpp
@@ -36,8 +36,7 @@ void PhaseButton::paint(QPainter *painter, const QStyleOptionGraphicsItem * /*op
     QRectF iconRect = boundingRect().adjusted(3, 3, -3, -3);
     QRectF translatedIconRect = painter->combinedTransform().mapRect(iconRect);
     qreal scaleFactor = translatedIconRect.width() / iconRect.width();
-    QPixmap iconPixmap =
-        PhasePixmapGenerator::generatePixmap(static_cast<int>(qRound(translatedIconRect.height())), name);
+    QPixmap iconPixmap = PhasePixmapGenerator::generatePixmap(qRound(translatedIconRect.height()), name);
 
     painter->setBrush(QColor(static_cast<int>(220 * (activeAnimationCounter / 10.0)),
                              static_cast<int>(220 * (activeAnimationCounter / 10.0)),
@@ -46,9 +45,8 @@ void PhaseButton::paint(QPainter *painter, const QStyleOptionGraphicsItem * /*op
     painter->drawRect(0, 0, static_cast<int>(width - 1), static_cast<int>(width - 1));
     painter->save();
     resetPainterTransform(painter);
-    painter->drawPixmap(iconPixmap.rect().translated(static_cast<int>(qRound(3 * scaleFactor)),
-                                                     static_cast<int>(qRound(3 * scaleFactor))),
-                        iconPixmap, iconPixmap.rect());
+    painter->drawPixmap(iconPixmap.rect().translated(qRound(3 * scaleFactor), qRound(3 * scaleFactor)), iconPixmap,
+                        iconPixmap.rect());
     painter->restore();
 
     painter->setBrush(QColor(0, 0, 0, static_cast<int>(255 * ((10 - activeAnimationCounter) / 15.0))));
